requirements.txt
faker
fastapi
uvicorn
fastapi_pagination
pytest
requests
sqlalchemy
pymysql
python-dotenv
cryptography
pandas